+++ src/main.ts
@@ -46,8 +46,11 @@ export default class FullStatisticsPlugin extends Plugin {
 			setFullVaultMetrics(this.vaultMetrics).
 			setExcludedFolders(this.settings.excludedFolders).
 			setOwnTags(this.settings.ownTags).
+			setOwnFolders(this.settings.ownFolders).
 			setSourceTags(this.settings.sourceTags).
+			setSourceFolders(this.settings.sourceFolders).
 			setConceptTags(this.settings.conceptTags).
+			setConceptFolders(this.settings.conceptFolders).
 			setInboxFolders(this.settings.inboxFolders).
 			setInboxReviewTags(this.settings.inboxReviewTags);
 
@@ -324,8 +327,11 @@ export default class FullStatisticsPlugin extends Plugin {
 		this.vaultMetricsCollector
 			.setExcludedFolders(this.settings.excludedFolders)
 			.setOwnTags(this.settings.ownTags)
+			.setOwnFolders(this.settings.ownFolders)
 			.setSourceTags(this.settings.sourceTags)
+			.setSourceFolders(this.settings.sourceFolders)
 			.setConceptTags(this.settings.conceptTags)
+			.setConceptFolders(this.settings.conceptFolders)
 			.setInboxFolders(this.settings.inboxFolders)
 			.setInboxReviewTags(this.settings.inboxReviewTags)
 			.restart();